test: widen fuzz fragmentation arg to uint16 for full seam coverage
FuzzProcessSingleResponse used a byte for the packet-payload size, which
capped every packet at 256 bytes. For streams longer than 256 bytes that
prevented the target from placing a seam past offset 256 or delivering the
stream in the largest legal packet layout. Widen the fuzz argument to uint16
so mutations can drive the seam across the full payload range supported by
frameReplyPackets (which still clamps to fuzzMaxPacketPayload).

Refs #418

Copilot-Session: c3e268ca-c6f4-4edb-826c-3a4018cf939c

+++ token_fuzz_test.go
@@ -559,23 +559,24 @@ func TestProcessSingleResponsePacketBoundary(t *testing.T) {
 // test-infrastructure layer.
 func FuzzProcessSingleResponse(f *testing.F) {
 	for _, seed := range fuzzResponseSeeds() {
-		f.Add(seed, byte(0))
-		f.Add(seed, byte(3))
+		f.Add(seed, uint16(0))
+		f.Add(seed, uint16(3))
 	}
 	// A couple of raw single-token seeds for extra coverage.
-	f.Add([]byte{byte(tokenColMetadata)}, byte(0))
-	f.Add([]byte{}, byte(0))
+	f.Add([]byte{byte(tokenColMetadata)}, uint16(0))
+	f.Add([]byte{}, uint16(0))
 
-	f.Fuzz(func(t *testing.T, stream []byte, frag byte) {
+	f.Fuzz(func(t *testing.T, stream []byte, frag uint16) {
 		// Bound input size to keep framing and allocations reasonable. A TDS
 		// packet length is a uint16, and the read buffer is 32 KiB, so very
 		// large inputs would either fail to frame or blow the buffer.
 		if len(stream) > 64*1024 {
 			t.Skip()
 		}
-		// Interpret the fuzzed byte as a packet payload size (1..256) so the
-		// engine can drive the seam to arbitrary offsets rather than a fixed
-		// set of even splits.
+		// Interpret the fuzzed value as a packet payload size (1..65536) so the
+		// engine can drive the seam to any offset across the full payload range
+		// that frameReplyPackets supports, not just the first 256 bytes.
+		// frameReplyPackets clamps chunk to fuzzMaxPacketPayload.
 		chunk := 1 + int(frag)
 		// The invariant: this must return normally (no panic escaping the
 		// parser's recover, no goroutine leak/deadlock). The returned values
